random placement works now
